backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
alembic>=1.7.1
pymysql>=1.0.2
python-dotenv>=0.19.0
opencv-python>=4.5.3
numpy>=1.21.2
tensorflow>=2.6.0
requests>=2.26.0  # For testing